+++ R/tm_g_scatterplotmatrix.R
@@ -151,11 +151,9 @@ srv_g_scatterplotmatrix <- function(input,
 
   init_chunks()
 
-  selector_list <- reactive({
-    data_extract_multiple_srv(data_extract = list(variables = variables), datasets = datasets)
-  })
+  selector_list <- data_extract_multiple_srv(data_extract = list(variables = variables), datasets = datasets)
 
-  merged_data <- data_merge_module_srv(
+  merged_data <- data_merge_srv(
     datasets = datasets,
     selector_list = selector_list
   )